pyproject.toml
[tool.poetry]
name = "faq-floripa"
version = "0.1.0"
description = ""
authors = ["Thiago Guimarães <thiagodsa@usp.br>"]
license = "MIT"

[tool.poetry.dependencies]
spacy = "^3.7.0"
pt_core_news_sm = {url = "https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.7.0/pt_core_news_sm-3.7.0.tar.gz"}
en_core_web_sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz"}
pydantic = "2.11.4"
python = "^3.9"
annotated-types = "0.7.0"
anyio = "4.9.0"
catalogue = "2.0.10"
certifi = "2025.4.26"
charset-normalizer = "3.4.2"
click = "8.1.8"
cloudpathlib = "0.21.1"
confection = "0.1.5"
cymem = "2.0.11"
distro = "1.9.0"
dotenv = "0.9.9"
exceptiongroup = "1.3.0"
faiss-cpu = "1.11.0"
filelock = "3.18.0"
fsspec = "2025.3.2"
groq = "0.25.0"
h11 = "0.16.0"
httpcore = "1.0.9"
httpx = "0.28.1"
huggingface-hub = "0.31.2"
idna = "3.10"
jinja2 = "3.1.6"
joblib = "1.5.0"
langcodes = "3.5.0"
langdetect = "1.0.9"
language-data = "1.3.0"
marisa-trie = "1.2.1"
markdown-it-py = "3.0.0"
markupsafe = "3.0.2"
mdurl = "0.1.2"
mpmath = "1.3.0"
murmurhash = "1.0.12"
nltk = "3.9.1"
nvidia-cublas-cu12 = "12.6.4.1"
nvidia-cuda-cupti-cu12 = "12.6.80"
nvidia-cuda-nvrtc-cu12 = "12.6.77"
nvidia-cuda-runtime-cu12 = "12.6.77"
nvidia-cudnn-cu12 = "9.5.1.17"
nvidia-cufft-cu12 = "11.3.0.4"
nvidia-cufile-cu12 = "1.11.1.6"
nvidia-curand-cu12 = "10.3.7.77"
nvidia-cusolver-cu12 = "11.7.1.2"
nvidia-cusparse-cu12 = "12.5.4.2"
nvidia-cusparselt-cu12 = "0.6.3"
nvidia-nccl-cu12 = "2.26.2"
nvidia-nvjitlink-cu12 = "12.6.85"
nvidia-nvtx-cu12 = "12.6.77"
packaging = "25.0"
pillow = "11.2.1"
preshed = "3.0.9"
pydantic-core = "2.33.2"
pygments = "2.19.1"
python-dotenv = "1.1.0"
python-telegram-bot = "22.1"
pyyaml = "6.0.2"
regex = "2024.11.6"
requests = "2.32.3"
rich = "14.0.0"
safetensors = "0.5.3"
scikit-learn = "1.6.1"
sentence-transformers = "4.1.0"
shellingham = "1.5.4"
six = "1.17.0"
smart-open = "7.1.0"
sniffio = "1.3.1"
spacy-legacy = "3.0.12"
spacy-loggers = "1.0.5"
sympy = "1.14.0"
threadpoolctl = "3.6.0"
tokenizers = "0.21.1"
torch = "2.7.0"
tqdm = "4.67.1"
transformers = "4.51.3"
triton = "3.3.0"
typer = "0.15.4"
typing-inspection = "0.4.0"
typing-extensions = "4.13.2"
urllib3 = "2.4.0"
wasabi = "1.1.3"
weasel = "0.4.1"
wrapt = "1.17.2"

